Revert "fix variable type"
This reverts commit 735d7b1376d52a3fe6f52f035ba64b6da19de757.

+++ evio-4.3.1/include/evioUtilTemplates.hxx
@@ -816,7 +816,7 @@ template <class Predicate> evioDOMNodeP evioDOMTree::findFirstNode(evioDOMNodeP 
 template <typename T> vector<T> *evioDOMTree::getVectorUnique(void) throw(evioException) {
 
   evioDOMNodeListP l = getNodeList(typeIs<T>());
-  size_t s = l->size();
+  int s = l->size();
 
   if(s==0) {
     return(NULL);
